Updated Model with better Lidar code
Lidar model now uses the robot heading as a parameter
Model handles annoying features such as the lidar's different coordinate system and that it is rotated by 180 degrees to the IMU
Produces obstacles array of detected points with their 'width' in a coordinate system relative to the robot's heading.
Potentially will change this so it produces "world" coordinates. However, this is potentially problematic since the coordinates are reliant on the estimation of the robot's pose, so if this drifts, the "world" coordinates will be incorrect. Also, should they be updated? IDK